+++ src/main/scala/org/openhorizon/exchangeapi/table/agreementbot/deploymentpolicy/AgbotBusinessPol.scala
@@ -2,5 +2,15 @@ package org.openhorizon.exchangeapi.table.agreementbot.deploymentpolicy
 
 final case class AgbotBusinessPol(businessPolOrgid: String,
                                   businessPol: String,
-                                  nodeOrgid: String,
-                                  lastUpdated: String)
+                                  // deployment_policy: String,
+                                  lastUpdated: String,
+                                  nodeOrgid: String) {
+  // Constructor
+  def this(agbotBusinessPolRow: AgbotBusinessPolRow) = {
+    this(businessPolOrgid = agbotBusinessPolRow.businessPolOrgid,
+         businessPol      = agbotBusinessPolRow.businessPol,
+         // deployment_policy = agbotBusinessPolRow.busPolId,
+         lastUpdated      = agbotBusinessPolRow.lastUpdated,
+         nodeOrgid        = agbotBusinessPolRow.nodeOrgid)
+  }
+}