+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [2.151.0](https://github.com/open-horizon/exchange-api/pull/826) - 2025-10-16
+- Fixed authorization for agreement bot deployment routes.
+- Improved debug logging for agreement bot routes.
+- Fixed agreement bot deployment routes missing from swagger documentation.
+- SBT 1.11.0 -> 1.11.7
+
 ## [2.150.0](https://github.com/open-horizon/exchange-api/pull/824) - 2025-09-25
 - Database foreign key cleanup. Removing a few duplicate keys, correcting cascade rules, and renaming a few more.
 - Improvements to debug logging.